refactor: drag mode 파일 정리
